// src/app/models/publicaciones.ts
export class Publicacion {
  idPost?: string;
  tipoPost: string
  categoriaPost: string
  estadoPost?: string;
  tituloPost?: string;
  autorNamePost?: string;
  viewsPost?: number;
  fechaPost?: Date;
  fechaInicioPost?:  Date;
  fechaFinPost?:  Date;
  descripcionPost?: string;
  horainicioPost?:string;
  horafinPost?:string;
  telPost?:string;
  lugarPost?:string;
  imagenPost?: any[];
  autorIdPost: string;
  autorImagenPost: string;
  docsPost?: any[];
  ytUrlPost?: string[];
  idGrupoPost?:string;
  comentarioPost?:string;
  nameGroupPost?:string;
  idUserAprobRech?:string;
  }